requirements.txt
python-telegram-bot
plexapi
jinja2